% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dfs.R
\name{get_dfs}
\alias{get_dfs}
\title{get_dfs}
\usage{
get_dfs(col)
}
\arguments{
\item{col}{column of the regression design matrix}
}
\value{
This function returns a numeric value indicating the degrees of freedom of a linear regression covariate.
}
\description{
Calculates the degrees of freedom for a covariate in a linear regression model
}
\examples{
data(mtcars)
mod = lm(mpg ~ cyl + disp, data = mtcars)
get_dfs(mod$model[["cyl"]])

}
