% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_anova.R
\name{my_anova}
\alias{my_anova}
\title{my_anova}
\usage{
my_anova(...)
}
\arguments{
\item{...}{one or more objects of class "lm"}
}
\value{
This function returns an object of class anova. These objects represent analysis-of-variance tables.
}
\description{
Perform ANOVA analyses on one or more linear regression models
}
\examples{
data(iris)
mod1 = lm(Sepal.Length ~ Sepal.Width, data = iris)
mod2 = lm(Sepal.Length ~ Sepal.Width + Petal.Length + Petal.Width, data = iris)
my_anova(mod2)
my_anova(mod1, mod2)

}
