% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_anova1.R
\name{my_anova1}
\alias{my_anova1}
\title{my_anova1}
\usage{
my_anova1(mod)
}
\arguments{
\item{mod}{an "lm" object with at least one covariate in the linear regression model}
}
\value{
This function returns an object of class anova. These objects represent analysis-of-variance tables.
}
\description{
Performs a sequential ANOVA analysis on one linear regression model.
}
\examples{
data(mtcars)
mod = lm(mpg ~ cyl + disp, data = mtcars)
my_anova1(mod)

}
