% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_anova1_intercept.R
\name{my_anova1_interceptonly}
\alias{my_anova1_interceptonly}
\title{my_anova1_interceptonly}
\usage{
my_anova1_interceptonly(mod)
}
\arguments{
\item{mod}{an "lm" object containing an intercept-only linear regression model}
}
\value{
This function returns an object of class anova. These objects represent analysis-of-variance tables.

For a complete breakdown of the table output, please see the myANOVA vignette.
}
\description{
Performs an ANOVA analysis on an intercept-only linear regression model.
}
\examples{
data(iris)
mod = lm(Sepal.Length ~ 1, data = iris)
my_anova1_interceptonly(mod)

}
