% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_anova1_nocoeffs.R
\name{my_anova1_nocoeffs}
\alias{my_anova1_nocoeffs}
\title{my_anova1_nocoeffs}
\usage{
my_anova1_nocoeffs(mod)
}
\arguments{
\item{mod}{an "lm" object with at least one covariate in the linear regression model}
}
\value{
This function returns an object of class anova. These objects represent analysis-of-variance tables.

For a complete breakdown of the table output, please see the myANOVA vignette.
}
\description{
Performs an ANOVA analysis on one no-coefficient linear regression model.
}
\examples{
data(iris)
mod = lm(Sepal.Length ~ -1, data = iris)
my_anova1_nocoeffs(mod)

}
