% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_anova2.R
\name{my_anova2}
\alias{my_anova2}
\title{my_anova2}
\usage{
my_anova2(mods)
}
\arguments{
\item{mods}{list containing at least 2 objects of type "lm"}
}
\value{
This function returns an object of class anova. These objects represent analysis-of-variance tables.
}
\description{
Perform ANOVA analyses on multiple linear regression models
}
\examples{
data(mtcars)
mod1 = lm(mpg ~ cyl, data = mtcars)
mod2 = lm(mpg ~ cyl + disp, data = mtcars)
my_anova2(mod1, mod2)

}
