% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_anova_intercept.R
\name{my_anova_intercept}
\alias{my_anova_intercept}
\title{my_anova_intercept}
\usage{
my_anova_intercept(mod)
}
\arguments{
\item{mod}{an "lm" object containing an intercept-only linear regression model}
}
\value{
This function returns an object of class anova. These objects represent analysis-of-variance tables.
}
\description{
Performs an ANOVA analysis on an intercept-only linear regression model.
}
\examples{
data(mtcars)
mod = lm(mpg ~ 1, data = mtcars)
my_anova_intercept(mod)

}
