% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/oneway.R
\docType{package}
\name{oneway.default}
\alias{oneway.default}
\alias{oneway.default-package}
\title{Oneway ANOVA table}
\arguments{
\item{z}{A list that contains groups.}
}
\value{
The list contains that ssb, ssw, dfw, dfb ...
}
\description{
Takes a list as an input and return the Some of Squares Within and Between , Degrees of Freedom
Between and Within, Group names and thier factors, Means values of each group.
}

