% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_anova.R
\name{my_anova}
\alias{my_anova}
\title{my_anova}
\usage{
my_anova(...)
}
\arguments{
\item{...}{One or more comma separated objects of class "lm". If only one lm object is provided, it must be a main effect model, not an interaction model.}
}
\value{
This function returns an object of class anova. These objects represent analysis-of-variance tables.

For a complete breakdown of the table output, please see the myANOVA vignette.
}
\description{
Perform ANOVA analyses on one or more linear regression models
}
\examples{
data(iris)
mod1 = lm(Sepal.Length ~ Sepal.Width, data = iris)
mod2 = lm(Sepal.Length ~ Sepal.Width + Petal.Length + Petal.Width, data = iris)
my_anova(mod2)
my_anova(mod1, mod2)

}
