% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_anova1.R
\name{my_anova1}
\alias{my_anova1}
\title{my_anova1}
\usage{
my_anova1(mod)
}
\arguments{
\item{mod}{an "lm" object with at least one covariate and an intercept in the linear regression model}
}
\value{
This function returns an object of class anova. These objects represent analysis-of-variance tables.

For a complete breakdown of the table output, please see the myANOVA vignette.
}
\description{
Performs a sequential ANOVA analysis on one linear regression model.
}
\examples{
data(iris)
mod = lm(Sepal.Length ~ Sepal.Width + Petal.Length + Petal.Width + Species, data = iris)
my_anova1(mod)

}
