% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_anova1_nointercept.R
\name{my_anova1_nointercept}
\alias{my_anova1_nointercept}
\title{my_anova1_nointercept}
\usage{
my_anova1_nointercept(mod)
}
\arguments{
\item{mod}{an "lm" object with no intercept and no interactions in the linear model}
}
\value{
This function returns an object of class anova. These objects represent analysis-of-variance tables.

For a complete breakdown of the table output, please see the myANOVA vignette.
}
\description{
Performs a sequential ANOVA analysis on one no-intercept main effect linear regression model.
}
\examples{
data(iris)
mod = lm(Sepal.Length ~ -1 + Sepal.Width + Species, data = iris)
my_anova1_nointercept(mod)

}
