% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_anova2.R
\name{my_anova2}
\alias{my_anova2}
\title{my_anova2}
\usage{
my_anova2(mods)
}
\arguments{
\item{mods}{list containing at least 2 objects of type "lm"}
}
\value{
This function returns an object of class anova. These objects represent analysis-of-variance tables.

For a complete breakdown of the table output, please see the myANOVA vignette.
}
\description{
Perform ANOVA analyses on multiple linear regression models
}
\examples{
data(iris)
mod1 = lm(Sepal.Length ~ Sepal.Width, data = iris)
mod2 = lm(Sepal.Length ~ Sepal.Width + Petal.Length + Petal.Width, data = iris)
mods = list(mod1, mod2)
my_anova2(mods)

}
